+++ doc/changelog.rst
@@ -3,7 +3,7 @@ Changelog
 
 v0.4 (*unreleased*)
 -------------------
-
+- don't detect comments ending with a colon as a block (:issue:`67`, :pull:`68`)
 
 v0.3 (04 November 2020)
 -----------------------